fix:aggregation bugs

+++ src/main/java/org/apache/iotdb/infludb/qp/logical/aggregation/AggregationValue.java
@@ -2,9 +2,9 @@ package org.apache.iotdb.infludb.qp.logical.aggregation;
 
 public class AggregationValue {
     private Object value;
-    private int timestamp;
+    private Long timestamp;
 
-    public AggregationValue(Object value, int timestamp) {
+    public AggregationValue(Object value, Long timestamp) {
         this.value = value;
         this.timestamp = timestamp;
     }
@@ -17,11 +17,11 @@ public class AggregationValue {
         this.value = value;
     }
 
-    public int getTimestamp() {
+    public Long getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(int timestamp) {
+    public void setTimestamp(Long timestamp) {
         this.timestamp = timestamp;
     }
 }
